refactor(indexer): nest replica state into sub-system owned by replica-log
Move catalogs, live-index, state, and services (indexer, compactor, tx-source, processor) into a nested integrant sub-system under :xtdb.indexer/replica-log.
This establishes ownership: the replica indexer owns its state, preparing for source-side state later.

The `into opts` pattern on expand-keys allows components to work in both contexts — ig/ref defaults resolve in the parent db-system, while literal values injected from the parent override them in the nested system.
