+++ core/src/main/clojure/xtdb/block_catalog.clj
@@ -6,9 +6,9 @@
   (:import (xtdb.block.proto Block TxKey)
            (xtdb.catalog BlockCatalog)))
 
-(defmethod ig/expand-key :xtdb/block-catalog [k {:keys [db-name]}]
-  {k {:db-name db-name
-      :buffer-pool (ig/ref :xtdb/buffer-pool)}})
+(defmethod ig/expand-key :xtdb/block-catalog [k opts]
+  {k (into {:buffer-pool (ig/ref :xtdb/buffer-pool)}
+           opts)})
 
 (defmethod ig/init-key :xtdb/block-catalog [_ {:keys [db-name buffer-pool]}]
   (BlockCatalog. db-name (BlockCatalog/getLatestBlock buffer-pool)))
