+++ core/src/main/clojure/xtdb/compactor.clj
@@ -30,12 +30,11 @@
 (defmethod ig/halt-key! :xtdb/compactor [_ compactor]
   (util/close compactor))
 
-(defmethod ig/expand-key ::for-db [k {:keys [base mode]}]
-  {k {:base base
-      :mode mode
-      :allocator (ig/ref :xtdb.db-catalog/allocator)
-      :storage (ig/ref :xtdb.db-catalog/storage)
-      :state (ig/ref :xtdb.db-catalog/state)}})
+(defmethod ig/expand-key ::for-db [k opts]
+  {k (into {:allocator (ig/ref :xtdb.db-catalog/allocator)
+            :storage (ig/ref :xtdb.db-catalog/storage)
+            :state (ig/ref :xtdb.db-catalog/state)}
+           opts)})
 
 (defmethod ig/init-key ::for-db [_ {{:keys [^Compactor compactor]} :base, :keys [allocator storage state ^Database$Mode mode]}]
   (if (= mode Database$Mode/READ_ONLY)
