+++ core/src/main/clojure/xtdb/log.clj
@@ -19,7 +19,7 @@
            (xtdb.arrow Relation Vector)
            xtdb.catalog.BlockCatalog
            (xtdb.database Database DatabaseStorage Database$Catalog Database$Mode)
-           xtdb.indexer.LogProcessor
+           xtdb.indexer.ReplicaLogProcessor
            xtdb.table.TableRef
            (xtdb.tx TxOp$DeleteDocs TxOp$EraseDocs TxOp$PatchDocs TxOp$PutDocs TxOp$Sql TxOpts TxWriter)
            (xtdb.tx_ops DeleteDocs EraseDocs PatchDocs PutDocs PutRel Sql SqlByteArgs)
@@ -242,30 +242,28 @@
                                                                                                              :system-time (some-> system-time time/expect-instant))))))]
         (MsgIdUtil/offsetToMsgId (.getEpoch log) (.getLogOffset message-meta))))))
 
-(defmethod ig/expand-key :xtdb.log/processor [k {:keys [base ^IndexerConfig indexer-conf ^Database$Mode mode db-catalog]}]
-  {k (cond-> {:base base
-              :allocator (ig/ref :xtdb.db-catalog/allocator)
-              :storage (ig/ref :xtdb.db-catalog/storage)
-              :state (ig/ref :xtdb.db-catalog/state)
-              :indexer (ig/ref :xtdb.indexer/for-db)
-              :compactor (ig/ref :xtdb.compactor/for-db)
-              :block-flush-duration (.getFlushDuration indexer-conf)
-              :skip-txs (.getSkipTxs indexer-conf)
-              :enabled? (.getEnabled indexer-conf)
-              :mode mode}
-       db-catalog (assoc :db-catalog db-catalog))})
+(defmethod ig/expand-key :xtdb.log/processor [k {:keys [^IndexerConfig indexer-conf] :as opts}]
+  {k (into {:allocator (ig/ref :xtdb.db-catalog/allocator)
+            :db-storage (ig/ref :xtdb.db-catalog/storage)
+            :db-state (ig/ref :xtdb.db-catalog/state)
+            :indexer (ig/ref :xtdb.indexer/for-db)
+            :compactor (ig/ref :xtdb.compactor/for-db)}
+           (assoc (dissoc opts :indexer-conf)
+                  :block-flush-duration (.getFlushDuration indexer-conf)
+                  :skip-txs (.getSkipTxs indexer-conf)
+                  :enabled? (.getEnabled indexer-conf)))})
 
 (defmethod ig/init-key :xtdb.log/processor [_ {{:keys [meter-registry]} :base
-                                               :keys [allocator ^DatabaseStorage storage state indexer compactor block-flush-duration skip-txs enabled? ^Database$Mode mode db-catalog]}]
+                                               :keys [allocator ^DatabaseStorage db-storage db-state indexer compactor block-flush-duration skip-txs enabled? ^Database$Mode mode db-catalog]}]
   (when enabled?
-    (let [lp (LogProcessor. allocator meter-registry
-                            storage state
+    (let [lp (ReplicaLogProcessor. allocator meter-registry
+                            db-storage db-state
                             indexer compactor block-flush-duration (set skip-txs)
                             (or mode Database$Mode/READ_WRITE)
                             1024
                             db-catalog)]
       {:processor lp
-       :subscription (.tailAll (.getSourceLog storage) lp (.getLatestProcessedOffset lp))})))
+       :subscription (.tailAll (.getSourceLog db-storage) lp (.getLatestProcessedOffset lp))})))
 
 (defmethod ig/resolve-key :xtdb.log/processor [_ {:keys [processor]}]
   processor)
